+++ sandbox/tests/anchors/anchors.ts
@@ -32,16 +32,10 @@ game.setAntialiasing(false);
 // center anchored actors
 var cl = new ex.Label('Centered', 0, 30);
 cl.textAlign = ex.TextAlign.Center;
-var ca1 = new ex.Actor(0, 0, 15, 15, ex.Color.Red);
-var ca2 = new ex.Actor(0, 0, 10, 10, ex.Color.Green);
-var ca3 = new ex.Actor(0, 0, 10, 10, ex.Color.Blue);
-var ca4 = new ex.Actor(0, 0, 20, 20);
-ca1.anchor.setTo(0.5, 0.5);
-ca2.anchor.setTo(0.5, 0.5);
-ca3.anchor.setTo(0.5, 0.5);
-ca4.anchor.setTo(0.5, 0.5);
-ca2.scale.setTo(2, 2);
-ca4.scale.setTo(2, 2);
+var ca1 = new ex.Actor({ x: 0, y: 0, width: 15, height: 15, color: ex.Color.Red, anchor: ex.Vector.Half });
+var ca2 = new ex.Actor({ x: 0, y: 0, width: 10 * 2, height: 10 * 2, color: ex.Color.Green, anchor: ex.Vector.Half });
+var ca3 = new ex.Actor({ x: 0, y: 0, width: 10, height: 10, color: ex.Color.Blue, anchor: ex.Vector.Half });
+var ca4 = new ex.Actor({ x: 0, y: 0, width: 20 * 2, height: 20 * 2, anchor: ex.Vector.Half });
 var heartSprite = heartTx.asSprite();
 heartSprite.scale.setTo(3, 3);
 ca4.addDrawing(heartSprite);
@@ -57,18 +51,39 @@ game.add(cl);
 // top left anchored actors
 var tll = new ex.Label('Top Left', -100, -60);
 tll.textAlign = ex.TextAlign.Center;
-var tla1 = new ex.Actor(-100, -100, 15, 15, ex.Color.Red);
-var tla2 = new ex.Actor(-100, -100, 10, 10, ex.Color.Green);
-var tla3 = new ex.Actor(-100, -100, 10, 10, ex.Color.Blue);
-var tla4 = new ex.Actor(-100, -100, 20, 20);
-tla1.anchor.setTo(0, 0);
-tla2.anchor.setTo(0, 0);
-tla3.anchor.setTo(0, 0);
-tla4.anchor.setTo(0, 0);
-tla2.scale.setTo(2, 2);
-tla4.scale.setTo(2, 2);
+var tla1 = new ex.Actor({
+  x: -100,
+  y: -100,
+  width: 15,
+  height: 15,
+  color: ex.Color.Red,
+  anchor: ex.Vector.Zero
+});
+var tla2 = new ex.Actor({
+  x: -100,
+  y: -100,
+  width: 10 * 2,
+  height: 10 * 2,
+  color: ex.Color.Green,
+  anchor: ex.Vector.Zero
+});
+var tla3 = new ex.Actor({
+  x: -100,
+  y: -100,
+  width: 10,
+  height: 10,
+  color: ex.Color.Blue,
+  anchor: ex.Vector.Zero
+});
+var tla4 = new ex.Actor({
+  x: -100,
+  y: -100,
+  width: 20 * 2,
+  height: 20 * 2,
+  anchor: ex.Vector.Zero
+});
 var heartSprite2 = heartTx.asSprite();
-heartSprite2.scale.setTo(2, 2);
+heartSprite2.scale.setTo(3, 3);
 tla4.addDrawing(heartSprite2);
 tla3.rotation = ex.Util.toRadians(45);
 
@@ -81,17 +96,11 @@ game.add(tll);
 // top right anchored actors
 var trl = new ex.Label('Top Right', 100, -60);
 trl.textAlign = ex.TextAlign.Center;
-var tra1 = new ex.Actor(100, -100, 15, 15, ex.Color.Red);
-var tra2 = new ex.Actor(100, -100, 10, 10, ex.Color.Green);
-var tra3 = new ex.Actor(100, -100, 10, 10, ex.Color.Blue);
-var tra4 = new ex.Actor(100, -100, 20, 20);
-tra1.anchor.setTo(1, 0);
-tra2.anchor.setTo(1, 0);
-tra3.anchor.setTo(1, 0);
-tra4.anchor.setTo(1, 0);
-tra2.scale.setTo(2, 2);
-tra4.scale.setTo(2, 2);
-tra4.addDrawing(heartTx);
+var tra1 = new ex.Actor({ x: 100, y: -100, width: 15, height: 15, color: ex.Color.Red, anchor: new ex.Vector(1, 0) });
+var tra2 = new ex.Actor({ x: 100, y: -100, width: 10 * 2, height: 10 * 2, color: ex.Color.Green, anchor: new ex.Vector(1, 0) });
+var tra3 = new ex.Actor({ x: 100, y: -100, width: 10, height: 10, color: ex.Color.Blue, anchor: new ex.Vector(1, 0) });
+var tra4 = new ex.Actor({ x: 100, y: -100, width: 20 * 2, height: 20 * 2, anchor: new ex.Vector(1, 0) });
+tra4.addDrawing(heartSprite2);
 tra3.rotation = ex.Util.toRadians(45);
 
 game.add(tra4);
@@ -103,17 +112,11 @@ game.add(trl);
 // bottom left anchored actors
 var bll = new ex.Label('Bottom Left', -100, 60);
 bll.textAlign = ex.TextAlign.Center;
-var bla1 = new ex.Actor(-100, 100, 15, 15, ex.Color.Red);
-var bla2 = new ex.Actor(-100, 100, 10, 10, ex.Color.Green);
-var bla3 = new ex.Actor(-100, 100, 10, 10, ex.Color.Blue);
-var bla4 = new ex.Actor(-100, 100, 20, 20);
-bla1.anchor.setTo(0, 1);
-bla2.anchor.setTo(0, 1);
-bla3.anchor.setTo(0, 1);
-bla4.anchor.setTo(0, 1);
-bla2.scale.setTo(2, 2);
-bla4.scale.setTo(2, 2);
-bla4.addDrawing(heartTx);
+var bla1 = new ex.Actor({ x: -100, y: 100, width: 15, height: 15, color: ex.Color.Red, anchor: new ex.Vector(0, 1) });
+var bla2 = new ex.Actor({ x: -100, y: 100, width: 10 * 2, height: 10 * 2, color: ex.Color.Green, anchor: new ex.Vector(0, 1) });
+var bla3 = new ex.Actor({ x: -100, y: 100, width: 10, height: 10, color: ex.Color.Blue, anchor: new ex.Vector(0, 1) });
+var bla4 = new ex.Actor({ x: -100, y: 100, width: 20 * 2, height: 20 * 2, anchor: new ex.Vector(0, 1) });
+bla4.addDrawing(heartSprite2);
 bla3.rotation = ex.Util.toRadians(45);
 
 game.add(bla4);
@@ -125,17 +128,11 @@ game.add(bll);
 // bottom right anchored actors
 var brl = new ex.Label('Bottom Right', 100, 60);
 brl.textAlign = ex.TextAlign.Center;
-var bra1 = new ex.Actor(100, 100, 15, 15, ex.Color.Red);
-var bra2 = new ex.Actor(100, 100, 10, 10, ex.Color.Green);
-var bra3 = new ex.Actor(100, 100, 10, 10, ex.Color.Blue);
-var bra4 = new ex.Actor(100, 100, 20, 20);
-bra1.anchor.setTo(1, 1);
-bra2.anchor.setTo(1, 1);
-bra3.anchor.setTo(1, 1);
-bra4.anchor.setTo(1, 1);
-bra2.scale.setTo(2, 2);
-bra4.scale.setTo(2, 2);
-bra4.addDrawing(heartTx);
+var bra1 = new ex.Actor({ x: 100, y: 100, width: 15, height: 15, color: ex.Color.Red, anchor: new ex.Vector(1, 1) });
+var bra2 = new ex.Actor({ x: 100, y: 100, width: 10 * 2, height: 10 * 2, color: ex.Color.Green, anchor: new ex.Vector(1, 1) });
+var bra3 = new ex.Actor({ x: 100, y: 100, width: 10, height: 10, color: ex.Color.Blue, anchor: new ex.Vector(1, 1) });
+var bra4 = new ex.Actor({ x: 100, y: 100, width: 20 * 2, height: 20 * 2, anchor: new ex.Vector(1, 1) });
+bra4.addDrawing(heartSprite2);
 bra3.rotation = ex.Util.toRadians(45);
 
 game.add(bra4);